Added scripts to initialize the database, and dump its contents directly
The create_test_data script contains hard-coded values to be used to play with
during development. Ultimately, this won't be used in the final product.

The dump_db.sh contains sqlite instructions to print everything from the
database. Because columns are selected explicitly, this will need to stay up to
date with changes to the schema (selecting '*' pulls in things like _etag,
_created, and _updated; which just clutter the screen).

The dump script will become overwhelming quickly once the data sizes become
realistic.