Added OFX Imports
The Open Financial Exchange format allows downloading of statement data for
accounts that support it (many do). These changes utilize the ofxclient package
to download the entries and add them to littlefinger's database.

This introduces a new configuration file, which is stored in INI format. This is
dictated by the ofxclient software. It seems possible to supply the information
manually, which would allow storing the information directly in our database,
but for now this is a straightforward way to configure the OFX components
separately. Note that passwords are stored in plaintext, which isn't great, but
will be the starting point.