Split API method into simple form and wrapped form, for easier debugging
By creating a routine that doesn't bother with Flask decorations, we can
directly invoke it in test scripts. This helps get things started, at least, and
theoretically would help make this function available for use elsewhere in the
application.

Of course, this is just a goofy test endpoint, it doesn't really have a coherent
purpose, and will probably be removed eventually.