Initial angular webapp specification
This is based on the angular-cli conventions. I deleted some of the things I
don't plan on using (like e2e with protractor), and didn't focus too much on the
unit tests yet. It's a starting point to get a basic functionality going, so I
can start to evolve the application's direction.